MIM-275 free talk on CASU channel + MIM-238 fetch unique action (#208)
* MIM-275 add free text on CASU channel

* make radio message input generic

* new radioMessageInput page

* MIM-238 Fetch unique action templates

* remove useless ActionTypes.

* Temporary fix for isChannelNewActivityDisabled. 

* Remove getAvailableActions

* Remove isChannelNewActivityDisabled

It was useful when we wanted to restrict the channel to one free message per channel. Now the restriction that an actor can play only one action on a time slot prevent (implemented in customCanConcurrencyWiseBePlayed in actionTemplateBase)

* removing template references first step

* removing template references continued. added canCancel functions

* removing template references continued. updated pages scripts

* added ignore DS_store files in zip creation script

* MIM-238 finishing touches

---------
