+++ gameModel/libs/ClientScript/UIfacade/actionFacade.ts
@@ -1,9 +1,11 @@
 /**
- * All UX interactions related to actions should live here
- * if any signature is modified make sure to report it in all page scripts
- * put minimal logic in here
+ * All UX interactions related to actions should live here.
+ * If any signature is modified make sure to report it in all page scripts.
+ * Put minimal logic in here.
  */
 
+import { IUniqueActionTemplates } from '../game/actionTemplatesData';
+import { ActionType, RadioType } from '../game/common/actionType';
 import { ActionBase } from '../game/common/actions/actionBase';
 import {
   ActionTemplateBase,
@@ -16,37 +18,83 @@ import {
   SendRadioMessageTemplate,
   SimFlag,
 } from '../game/common/actions/actionTemplateBase';
-import { ActorId, TemplateId, TemplateRef } from '../game/common/baseTypes';
-import { ActionCreationEvent } from '../game/common/events/eventTypes';
+import { ActorId, TemplateId } from '../game/common/baseTypes';
 import {
   buildAndLaunchActionCancellation,
   buildAndLaunchActionFromTemplate,
   fetchAvailableActions,
   getCurrentState,
+  getUniqueActionTemplates,
 } from '../game/mainSimulationLogic';
-import { ActionType } from '../game/common/actionType';
+import { getTypedInterfaceState } from '../gameInterface/interfaceState';
+import { canPlanAction, isPlannedAction } from '../gameInterface/main';
+
+// used in page 45 (actionStandardList)
+export function getAvailableActionTemplates(
+  actionType: ActionType = ActionType.ACTION
+): ActionTemplateBase[] {
+  const currentActorUid = getTypedInterfaceState().currentActorUid;
+  if (currentActorUid) {
+    return fetchAvailableActions(currentActorUid, actionType);
+  }
+
+  return [];
+}
+
+export function isAvailable(template: ActionTemplateBase): boolean {
+  const currentActorUid = getTypedInterfaceState().currentActorUid;
+  if (template && currentActorUid) {
+    const state = getCurrentState();
+    const actor = state.getActorById(currentActorUid);
+    if (actor) {
+      return template.isAvailable(state, actor);
+    }
+  }
+  return false;
+}
+
+/**
+ * @param template
+ * @returns true if the action can be played or is currently planned, thus can be cancelled
+ */
+export function canCancel(template: ActionTemplateBase): boolean {
+  return isAvailable(template) && isPlannedAction(template.Uid);
+}
+
+/**
+ * @param template
+ * @returns true if an action can be planned by the current actor
+ * or that the current actor can cancel an action based on this template
+ */
+export function canPlanOrCancel(template: ActionTemplateBase): boolean {
+  return canPlanAction() || canCancel(template);
+}
+
+export function uniqueActionTemplates(): IUniqueActionTemplates {
+  return getUniqueActionTemplates();
+}
 
 // TODO there might be specific local UI state to add in there (like a selected position or geometry)
 /**
  *
- * @param actionTemplateId The template to instanciate
+ * @param actionTemplate The template to instanciate
  * @param selectedActor The actor the plans the action and will be its owner
  * @param params The additional optional parameters, related to the chosen action template
  * @returns a promise
  */
 export async function planAction(
-  actionTemplateId: TemplateRef,
+  actionTemplate: ActionTemplateBase,
   selectedActor: ActorId,
   params?: any
 ): Promise<IManagedResponse | undefined> {
-  return await buildAndLaunchActionFromTemplate(actionTemplateId, selectedActor, params);
+  return await buildAndLaunchActionFromTemplate(actionTemplate, selectedActor, params);
 }
 
 // TODO Maybe ensure only owning actor can cancel actions
 /**
  *
- * @param actionId The action to cancel
  * @param selectedActor The actor that cancels the action
+ * @param templateId The template of the action to cancel
  * @returns
  */
 export async function cancelAction(
@@ -54,17 +102,6 @@ export async function cancelAction(
   templateId: TemplateId
 ): Promise<IManagedResponse | undefined> {
   return await buildAndLaunchActionCancellation(selectedActor, templateId);
-}
-
-/**
- * @param actorId
- * @returns a list of actions that the current actor can undertake
- */
-export function getAvailableActions(
-  actorId: ActorId,
-  actionType: ActionType = ActionType.ACTION
-): ActionTemplateBase[] {
-  return fetchAvailableActions(actorId, actionType);
 }
 
 /**
@@ -78,85 +115,36 @@ export function getAllCancelledActions(): Readonly<ActionBase[]> {
   return getCurrentState().getAllCancelledActions();
 }
 
-/**
- * @returns Template of given action Uid
- */
-export function getActionTemplate(
-  id: number,
-  actionType: ActionType = ActionType.ACTION
-): ActionTemplateBase<ActionBase, ActionCreationEvent, unknown> | undefined {
-  return getAvailableActions(Context.interfaceState.state.currentActorUid, actionType).find(
-    t => t.Uid === id
-  );
-}
-
-/**
- * @param id Uid of given action
- */
-
-export function isFixedMapEntityTemplate(id: number): boolean {
-  const template = getAvailableActions(Context.interfaceState.state.currentActorUid).find(
-    t => t.Uid === id
-  );
+export function isFixedMapEntityTemplate(template: ActionTemplateBase | undefined): boolean {
   return template instanceof SelectionFixedMapEntityTemplate;
 }
 
-/**
- * @param id Uid of given action
- */
-export function isCasuMessageActionTemplate(id: number): boolean {
-  const template = getAvailableActions(
-    Context.interfaceState.state.currentActorUid,
-    ActionType.CASU_RADIO
-  ).find(t => t.Uid === id);
+export function isCasuMessageActionTemplate(template: ActionTemplateBase | undefined): boolean {
   return template instanceof CasuMessageTemplate;
 }
 
-/**
- * @param id Uid of given action
- */
-export function isRadioActionTemplate(id: number): boolean {
-  const template = getAvailableActions(
-    Context.interfaceState.state.currentActorUid,
-    ActionType.ACTORS_RADIO
-  ).find(t => t.Uid === id);
-  return template instanceof SendRadioMessageTemplate;
+export function isRadioActionTemplate(
+  template: ActionTemplateBase | undefined,
+  radioChannel: RadioType
+): boolean {
+  return template instanceof SendRadioMessageTemplate && template.radioChannel === radioChannel;
 }
 
-/**
- * @param id Uid of given action template
- */
-export function isMoveResourcesAssignTaskActionTemplate(id: number): boolean {
-  const template = getAvailableActions(
-    Context.interfaceState.state.currentActorUid,
-    ActionType.ALLOCATE_RESOURCES
-  ).find(t => t.Uid === id);
+export function isMoveResourcesAssignTaskActionTemplate(
+  template: ActionTemplateBase | undefined
+): boolean {
   return template instanceof MoveResourcesAssignTaskActionTemplate;
 }
 
-/**
- * @param id Uid of given action template
- */
-export function isMoveActorActionTemplate(id: number): boolean {
-  const template = getAvailableActions(Context.interfaceState.state.currentActorUid).find(
-    t => t.Uid === id
-  );
+export function isMoveActorActionTemplate(template: ActionTemplateBase | undefined): boolean {
   return template instanceof MoveActorActionTemplate;
 }
 
-export function isEvacuationActionTemplate(id: number): boolean {
-  const template = getAvailableActions(
-    Context.interfaceState.state.currentActorUid,
-    ActionType.EVASAN_RADIO
-  ).find(t => t.Uid === id);
+export function isEvacuationActionTemplate(template: ActionTemplateBase | undefined): boolean {
   return template instanceof EvacuationActionTemplate;
 }
 
-export function isPretriageReportTemplate(id: number): boolean {
-  const template = getAvailableActions(
-    Context.interfaceState.state.currentActorUid,
-    ActionType.PRETRIAGE_REPORT
-  ).find(t => t.Uid === id);
+export function isPretriageReportTemplate(template: ActionTemplateBase | undefined): boolean {
   return template instanceof PretriageReportTemplate;
 }
