+++ gameModel/libs/ClientScript/game/common/actionType.ts
@@ -4,9 +4,6 @@ export enum ActionType {
   ACTORS_RADIO = 'ACTORS_RADIO',
   RESOURCES_RADIO = 'RESOURCES_RADIO',
   EVASAN_RADIO = 'EVASAN_RADIO',
-  ALLOCATE_RESOURCES = 'ALLOCATE_RESOURCES',
-  PRETRIAGE_REPORT = 'PRETRIAGE_REPORT',
-  ACTIVATE_RADIO_CHANNELS = 'ACTIVATE_RADIO_CHANNELS',
 }
 
 export type RadioType =
