+++ gameModel/libs/ClientScript/game/common/baseTypes.ts
@@ -15,8 +15,6 @@ export type LocalEventId = number;
 
 export type GlobalEventId = number;
 
-export type TemplateRef = string;
-
 export type ActorId = number;
 
 export type TemplateId = number;
