+++ gameModel/libs/ClientScript/tools/logger.ts
@@ -25,11 +25,10 @@ export const mainSimMapLogger = Helpers.getLogger('mainSim.map');
 localEventManagerLogger.setLevel('INFO');
 export const mainSimLogger = Helpers.getLogger('main.simulation.logic');
 export const mainSimStateLogger = Helpers.getLogger('main.simulation.state');
+export const actionLogger = Helpers.getLogger('action');
 export const taskLogger = Helpers.getLogger('task');
 export const resourceLogger = Helpers.getLogger('resource');
 export const timeLogger = Helpers.getLogger('time');
 export const debugFacadeLogger = Helpers.getLogger('debug');
 mainSimLogger.setLevel('INFO');
 mainSimStateLogger.setLevel('DEBUG');
-taskLogger.setLevel('DEBUG');
-resourceLogger.setLevel('DEBUG');
