+++ gameModel/pages/45.json
@@ -172,7 +172,7 @@
                     "localScriptEval" : {
                       "script" : {
                         "@class" : "ClientScript",
-                        "content" : "import { runActionButton } from \"./gameInterface/actionsButtonLogic\";\nrunActionButton();",
+                        "content" : "import { runActionButton } from \"./gameInterface/actionsButtonLogic\";\nrunActionButton(Context.action);",
                         "language" : "TypeScript"
                       }
                     },
@@ -220,7 +220,7 @@
         "exposeAs" : "action",
         "getItemsFn" : {
           "@class" : "Script",
-          "content" : "import { getAvailableActions } from \"./UIfacade/actionFacade\";\ngetAvailableActions(Context.interfaceState.state.currentActorUid);",
+          "content" : "import { ActionType } from \"./game/common/actionType\";\nimport { getAvailableActionTemplates } from \"./UIfacade/actionFacade\";\ngetAvailableActionTemplates(ActionType.ACTION);",
           "language" : "TypeScript"
         },
         "id" : "",
