+++ gameModel/pages/actionRadioChannelActivation.json
@@ -9,7 +9,7 @@
         "style" : { },
         "text" : {
           "@class" : "Script",
-          "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\ngetRadioChannelsActivationTemplate()?.getDescription();",
+          "content" : "import { uniqueActionTemplates } from \"./UIfacade/actionFacade\";\nuniqueActionTemplates().ActivateRadioSchemaActionTemplate.getDescription();",
           "language" : "TypeScript"
         }
       },
@@ -22,23 +22,23 @@
           "className" : "abort",
           "condition" : {
             "@class" : "Script",
-            "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\nimport { isPlannedAction } from \"./gameInterface/main\";\nconst action = getRadioChannelsActivationTemplate();\n!!action && isPlannedAction(action.Uid);",
+            "content" : "import { canCancel, uniqueActionTemplates } from \"./UIfacade/actionFacade\";\ncanCancel(uniqueActionTemplates().ActivateRadioSchemaActionTemplate);",
             "language" : "typescript"
           }
         } ],
         "disableIf" : {
           "@class" : "Script",
-          "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\nimport { canPlanAction, isPlannedAction } from \"./gameInterface/main\";\nconst action =  getRadioChannelsActivationTemplate();\n  if (!action) return true;\nif (canPlanAction()) return false;\n!isPlannedAction(action.Uid);",
+          "content" : "import { canPlanOrCancel, isAvailable, uniqueActionTemplates } from \"./UIfacade/actionFacade\";\nconst action = uniqueActionTemplates().ActivateRadioSchemaActionTemplate;\nif(!isAvailable(action)){\n  return true;\n}\n!canPlanOrCancel(action);",
           "language" : "typescript"
         },
         "hideIf" : {
           "@class" : "Script",
-          "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\n!getRadioChannelsActivationTemplate();",
+          "content" : "import { isAvailable, uniqueActionTemplates } from \"./UIfacade/actionFacade\";\n!isAvailable(uniqueActionTemplates().ActivateRadioSchemaActionTemplate);",
           "language" : "typescript"
         },
         "label" : {
           "@class" : "Script",
-          "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\nimport { isPlannedAction } from \"./gameInterface/main\";\nimport { getTranslation } from \"./tools/translation\";\nconst action =  getRadioChannelsActivationTemplate();\n  if (!action) {\n    return undefined;\n  }\nconst isPlanned = isPlannedAction(action.Uid)\nisPlanned ? getTranslation('mainSim-interface', 'cancel') : getTranslation(`mainSim-interface`, `send-one-minute`);",
+          "content" : "import { uniqueActionTemplates } from \"./UIfacade/actionFacade\";\nimport { isPlannedAction } from \"./gameInterface/main\";\nimport { getTranslation } from \"./tools/translation\";\nconst isPlanned = isPlannedAction(uniqueActionTemplates().ActivateRadioSchemaActionTemplate.Uid);\ngetTranslation('mainSim-interface', isPlanned ? 'cancel' : 'send-one-minute');",
           "language" : "TypeScript"
         },
         "layoutStyle" : {
@@ -48,7 +48,7 @@
         "localScriptEval" : {
           "script" : {
             "@class" : "ClientScript",
-            "content" : "import { getRadioChannelsActivationTemplate } from \"./game/common/actions/actionLogic\";\nimport { runActionButton } from \"./gameInterface/actionsButtonLogic\";\n\nconst action =  getRadioChannelsActivationTemplate();\nrunActionButton(action);",
+            "content" : "import { uniqueActionTemplates } from \"./UIfacade/actionFacade\";\nimport { runActionButton } from \"./gameInterface/actionsButtonLogic\";\nrunActionButton(uniqueActionTemplates().ActivateRadioSchemaActionTemplate);",
             "language" : "TypeScript"
           }
         },
@@ -57,6 +57,11 @@
       },
       "type" : "Button"
     } ],
+    "hideIf" : {
+      "@class" : "Script",
+      "content" : "import { isAvailable, uniqueActionTemplates } from \"./UIfacade/actionFacade\";\n!isAvailable(uniqueActionTemplates().ActivateRadioSchemaActionTemplate);",
+      "language" : "typescript"
+    },
     "layout" : {
       "flexDirection" : "column"
     },
